moved git classes out of provider for cleanliness
