+++ ludvig/providers/_git/__init__.py
@@ -1,0 +1,7 @@
+from ._definitions import (
+    GitPack,
+    GitPackIndex,
+    GitTreeItem,
+    GitMainIndex,
+    GitObjectType,
+)
